+++ packages/react/src/Active.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import invariant from 'invariant';
+import { CuriConfig, AnyResponse } from '@curi/core';
 
 export interface ActiveProps {
   children: any;
@@ -8,11 +9,13 @@ export interface ActiveProps {
   params?: object;
   partial?: boolean;
   merge: (props: object) => object;
+  curi?: CuriConfig;
+  response?: AnyResponse;
 }
 
 class Active extends React.Component<ActiveProps, {}> {
   static contextTypes = {
-    curi: PropTypes.object.isRequired,
+    curi: PropTypes.object,
     curiResponse: PropTypes.object
   };
 
@@ -25,20 +28,22 @@ class Active extends React.Component<ActiveProps, {}> {
   }
 
   verifyActiveAddon() {
+    const curi = this.props.curi || this.context.curi;
     invariant(
-      this.context.curi.addons.active,
+      curi.addons.active,
       'You are attempting to use the "active" prop, but have not included the "active" ' +
         'addon (curi-addon-active) in your Curi configuration object.'
     );
   }
 
   render() {
-    const { curi, curiResponse } = this.context;
+    const curi = this.props.curi || this.context.curi;
+    const response = this.props.response || this.context.curiResponse;
     const { merge, partial, name, params, children } = this.props;
 
     // need to make a copy
     let childProps = { ...children.props };
-    const isActive = curi.addons.active(name, curiResponse, params, partial);
+    const isActive = curi.addons.active(name, response, params, partial);
     if (isActive) {
       childProps = merge(childProps);
     }