+++ packages/react/src/Link.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import invariant from 'invariant';
 import { CuriContext } from './interface';
+import { CuriConfig, AnyResponse } from '@curi/core';
 
 const canNavigate = (event: React.MouseEvent<HTMLElement>) => {
   return (
@@ -16,7 +17,8 @@ export interface ActiveLink {
   partial?: boolean;
 }
 
-export interface LinkProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
+export interface LinkProps
+  extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
   to: string;
   params?: object;
   details?: object;
@@ -24,6 +26,8 @@ export interface LinkProps extends React.AnchorHTMLAttributes<HTMLAnchorElement>
   active?: ActiveLink;
   anchor?: React.ComponentClass | React.StatelessComponent;
   target?: string;
+  curi?: CuriConfig;
+  response?: AnyResponse;
 }
 
 export interface LinkState {
@@ -32,7 +36,7 @@ export interface LinkState {
 
 class Link extends React.Component<LinkProps, LinkState> {
   static contextTypes = {
-    curi: PropTypes.object.isRequired,
+    curi: PropTypes.object,
     curiResponse: PropTypes.object
   };
 
@@ -43,7 +47,7 @@ class Link extends React.Component<LinkProps, LinkState> {
 
     if (canNavigate(event) && !this.props.target) {
       event.preventDefault();
-      const { curi } = this.context;
+      const curi = this.props.curi || this.context.curi;
       const { pathname } = this.state;
       const { to, params, details = {} } = this.props;
       const location = { pathname, ...details };
@@ -53,7 +57,7 @@ class Link extends React.Component<LinkProps, LinkState> {
 
   createPathname(props: LinkProps, context: CuriContext) {
     const { to, params } = props;
-    const { curi } = context;
+    const curi = props.curi || context.curi;
     const pathname =
       to != null
         ? curi.addons.pathname(to, params)
@@ -78,8 +82,9 @@ class Link extends React.Component<LinkProps, LinkState> {
   }
 
   verifyActiveAddon() {
+    const curi = this.props.curi || this.context.curi;
     invariant(
-      this.context.curi.addons.active,
+      curi.addons.active,
       'You are attempting to use the "active" prop, but have not included the "active" ' +
         'addon (curi-addon-active) in your Curi configuration object.'
     );
@@ -95,12 +100,16 @@ class Link extends React.Component<LinkProps, LinkState> {
       anchor,
       ...rest
     } = this.props;
-    const { curi, curiResponse } = this.context;
+    const curi = this.props.curi || this.context.curi;
+    const response = this.props.response || this.context.curiResponse;
     let anchorProps = rest;
-    const Anchor: React.ComponentClass | React.StatelessComponent | string = anchor ? anchor : 'a';
+    const Anchor:
+      | React.ComponentClass
+      | React.StatelessComponent
+      | string = anchor ? anchor : 'a';
     if (active) {
       const { partial, merge } = active;
-      const isActive = curi.addons.active(to, curiResponse, params, partial);
+      const isActive = curi.addons.active(to, response, params, partial);
       if (isActive) {
         anchorProps = merge(anchorProps);
       }